Fix most type hints

+++ torchgeo/datasets/cowc.py
@@ -144,7 +144,8 @@ class COWC(NonGeoDataset, abc.ABC):
         filename = os.path.join(self.root, self.images[index])
         with Image.open(filename) as img:
             array: "np.typing.NDArray[np.int_]" = np.array(img)
-            tensor = torch.from_numpy(array).float()
+            tensor = torch.from_numpy(array)
+            tensor = tensor.float()
             # Convert from HxWxC to CxHxW
             tensor = tensor.permute((2, 0, 1))
             return tensor
@@ -159,7 +160,8 @@ class COWC(NonGeoDataset, abc.ABC):
             the target
         """
         target = int(self.targets[index])
-        tensor = torch.tensor(target).float()
+        tensor = torch.tensor(target)
+        tensor = tensor.float()
         return tensor
 
     def _check_integrity(self) -> bool:
